fixing logging for fallback

+++ baselines/logger.py
@@ -100,9 +100,14 @@ class JSONOutputFormat(KVWriter):
         self.file.close()
 
 class CSVOutputFormat(KVWriter):
-    def __init__(self, filename):
-        self.file = open(filename, 'w+t')
-        self.keys = []
+    def __init__(self, filename, append=False):
+        if append and osp.exists(filename):
+            self.file = open(filename, 'r+t')
+            self.keys = self.file.readline().rstrip('\n').split(',')
+            self.file.seek(0, 2)  # seek to end
+        else:
+            self.file = open(filename, 'w+t')
+            self.keys = []
         self.sep = ','
 
     def writekvs(self, kvs):
@@ -171,7 +176,7 @@ class TensorBoardOutputFormat(KVWriter):
             self.writer.Close()
             self.writer = None
 
-def make_output_format(format, ev_dir, log_suffix=''):
+def make_output_format(format, ev_dir, log_suffix='', append=False):
     os.makedirs(ev_dir, exist_ok=True)
     if format == 'stdout':
         return HumanOutputFormat(sys.stdout)
@@ -180,7 +185,7 @@ def make_output_format(format, ev_dir, log_suffix=''):
     elif format == 'json':
         return JSONOutputFormat(osp.join(ev_dir, 'progress%s.json' % log_suffix))
     elif format == 'csv':
-        return CSVOutputFormat(osp.join(ev_dir, 'progress%s.csv' % log_suffix))
+        return CSVOutputFormat(osp.join(ev_dir, 'progress%s.csv' % log_suffix), append=append)
     elif format == 'tensorboard':
         return TensorBoardOutputFormat(osp.join(ev_dir, 'tb%s' % log_suffix))
     else:
@@ -369,7 +374,7 @@ def get_rank_without_mpi_import():
     return 0
 
 
-def configure(dir=None, format_strs=None, comm=None, log_suffix=''):
+def configure(dir=None, format_strs=None, comm=None, log_suffix='', append=False):
     """
     If comm is provided, average all numerical stats across that comm
     """
@@ -392,7 +397,7 @@ def configure(dir=None, format_strs=None, comm=None, log_suffix=''):
         else:
             format_strs = os.getenv('OPENAI_LOG_FORMAT_MPI', 'log').split(',')
     format_strs = filter(None, format_strs)
-    output_formats = [make_output_format(f, dir, log_suffix) for f in format_strs]
+    output_formats = [make_output_format(f, dir, log_suffix, append=append) for f in format_strs]
 
     Logger.CURRENT = Logger(dir=dir, output_formats=output_formats, comm=comm)
     if output_formats:
